+++ Sample/source/base64.h
@@ -88,11 +88,11 @@
 extern "C" {
 #endif
 
-int Base64encode_len(int len);
-int Base64encode(char * coded_dst, const char *plain_src,int len_plain_src);
+size_t Base64encode_len(size_t len);
+size_t Base64encode(char * coded_dst, const char *plain_src, size_t len_plain_src);
 
-int Base64decode_len(const char * coded_src);
-int Base64decode(char * plain_dst, const char *coded_src);
+size_t Base64decode_len(const char * coded_src);
+size_t Base64decode(char * plain_dst, const char *coded_src);
 
 #ifdef __cplusplus
 }
